main.py: Fix voice not updating in certain cases
Fixed perms for a member in the voicechat text channel not updating in
certain cases. These cases were:

 - The member joined another voice channel instead of leaving
 - The member was already in a voice channel, but then joined the
   General voice channel

+++ main.py
@@ -1292,17 +1292,17 @@ async def on_voice_state_update(before, after):
 	if after.server.id != productionserver:
 		return
 
-	if before.voice.voice_channel == None:
-		if after.voice.voice_channel.id == '153368831329435648':
-			# JOINED the voice channel.
-			overwrite = discord.PermissionOverwrite()
-			overwrite.read_messages = True
-			await client.edit_channel_permissions(voicetextchannel, after, overwrite)
+	if after.voice.voice_channel != None \
+	and after.voice.voice_channel.id == '153368831329435648':
+		# JOINED the voice channel.
+		overwrite = discord.PermissionOverwrite()
+		overwrite.read_messages = True
+		await client.edit_channel_permissions(voicetextchannel, after, overwrite)
 
-	elif after.voice.voice_channel == None:
-		if before.voice.voice_channel.id == '153368831329435648':
-			# LEFT the voice channel.
-			await client.delete_channel_permissions(voicetextchannel, after)
+	if before.voice.voice_channel != None \
+	and before.voice.voice_channel.id == '153368831329435648':
+		# LEFT the voice channel.
+		await client.delete_channel_permissions(voicetextchannel, after)
 
 @client.event
 async def on_channel_create(c):
